Move on-disk workflow representation in a sub-module
